move mapSettledWithConcurrency to utils.ts
