+++ src/utils.ts
@@ -635,6 +635,36 @@ function zipObjectUnless<T extends Key, U>(keys: T[], values: U[], unless: (valu
   return map;
 }
 
+async function mapSettledWithConcurrency<T, R>(
+  values: readonly T[],
+  concurrency: number,
+  mapper: (value: T, index: number) => Promise<R>,
+): Promise<PromiseSettledResult<R>[]> {
+  const results = new Array<PromiseSettledResult<R>>(values.length);
+  let nextIndex = 0;
+
+  async function runNext(): Promise<void> {
+    while (true) {
+      const index = nextIndex++;
+      if (index >= values.length) return;
+
+      try {
+        results[index] = {
+          status: "fulfilled",
+          value: await mapper(values[index], index),
+        };
+      } catch (reason) {
+        results[index] = { status: "rejected", reason };
+      }
+    }
+  }
+
+  const runnersCount = Math.min(values.length, Math.max(1, Math.floor(concurrency)));
+  await Promise.all(Array.from({ length: runnersCount }, runNext));
+
+  return results;
+}
+
 export {
   castArray,
   fastJoinedPath,
@@ -671,6 +701,7 @@ export {
   isString,
   isTruthy,
   isUndefined,
+  mapSettledWithConcurrency,
   memoize,
   negate,
   noop,
